clangd

+++ src/core/CutsceneRunner.cpp
@@ -1,7 +1,6 @@
 #include "CutsceneRunner.h"
 #include "StringUtils.h"
 #include <fstream>
-#include <sstream>
 #include <iostream>
 #include <set>
 
@@ -9,11 +8,9 @@ using StringUtils::parseDirection;
 using StringUtils::splitPipe;
 using StringUtils::splitSemicolon;
 
-bool CutsceneRunner::load(const std::string &path)
-{
+bool CutsceneRunner::load(const std::string &path) {
     std::ifstream file(path);
-    if (!file.is_open())
-    {
+    if (!file.is_open()) {
         std::cerr << "CutsceneRunner: cannot open " << path << "\n";
         return false;
     }
@@ -21,8 +18,7 @@ bool CutsceneRunner::load(const std::string &path)
     cutscene.steps.clear();
     cutscene.id.clear();
 
-    enum class Section
-    {
+    enum class Section {
         none,
         header,
         steps,
@@ -30,8 +26,7 @@ bool CutsceneRunner::load(const std::string &path)
     Section section = Section::none;
 
     std::string line;
-    while (std::getline(file, line))
-    {
+    while (std::getline(file, line)) {
         // Trim trailing whitespace
         while (!line.empty() && (line.back() == '\r' || line.back() == ' '))
             line.pop_back();
@@ -39,25 +34,20 @@ bool CutsceneRunner::load(const std::string &path)
         if (line.empty() || line[0] == '#')
             continue;
 
-        if (line == "[header]")
-        {
+        if (line == "[header]") {
             section = Section::header;
             continue;
         }
-        if (line == "[steps]")
-        {
+        if (line == "[steps]") {
             section = Section::steps;
             continue;
         }
 
-        if (section == Section::header)
-        {
+        if (section == Section::header) {
             auto parts = splitPipe(line);
             if (parts.size() >= 2 && parts[0] == "id")
                 cutscene.id = parts[1];
-        }
-        else if (section == Section::steps)
-        {
+        } else if (section == Section::steps) {
             auto parts = splitPipe(line);
             if (parts.empty())
                 continue;
@@ -65,57 +55,38 @@ bool CutsceneRunner::load(const std::string &path)
             const std::string &cmd = parts[0];
             CutsceneStep step{};
 
-            if (cmd == "move" && parts.size() >= 4)
-            {
+            if (cmd == "move" && parts.size() >= 4) {
                 step.type = CutsceneStep::Type::move;
                 step.target = parts[1];
                 step.x = std::stoi(parts[2]);
                 step.y = std::stoi(parts[3]);
-            }
-            else if (cmd == "walk" && parts.size() >= 3)
-            {
+            } else if (cmd == "walk" && parts.size() >= 3) {
                 step.type = CutsceneStep::Type::walk;
                 step.target = parts[1];
                 step.direction = parseDirection(parts[2]);
-            }
-            else if (cmd == "face" && parts.size() >= 3)
-            {
+            } else if (cmd == "face" && parts.size() >= 3) {
                 step.type = CutsceneStep::Type::face;
                 step.target = parts[1];
                 step.direction = parseDirection(parts[2]);
-            }
-            else if (cmd == "say" && parts.size() >= 3)
-            {
+            } else if (cmd == "say" && parts.size() >= 3) {
                 step.type = CutsceneStep::Type::say;
                 step.speaker = parts[1];
                 step.lines = splitSemicolon(parts[2]);
-            }
-            else if (cmd == "wait" && parts.size() >= 2)
-            {
+            } else if (cmd == "wait" && parts.size() >= 2) {
                 step.type = CutsceneStep::Type::wait;
                 step.frames = std::stoi(parts[1]);
-            }
-            else if (cmd == "sync")
-            {
+            } else if (cmd == "sync") {
                 step.type = CutsceneStep::Type::sync;
-            }
-            else if (cmd == "flag" && parts.size() >= 2)
-            {
+            } else if (cmd == "flag" && parts.size() >= 2) {
                 step.type = CutsceneStep::Type::flag;
                 step.flagName = parts[1];
-            }
-            else if (cmd == "hide" && parts.size() >= 2)
-            {
+            } else if (cmd == "hide" && parts.size() >= 2) {
                 step.type = CutsceneStep::Type::hide;
                 step.target = parts[1];
-            }
-            else if (cmd == "show" && parts.size() >= 2)
-            {
+            } else if (cmd == "show" && parts.size() >= 2) {
                 step.type = CutsceneStep::Type::show;
                 step.target = parts[1];
-            }
-            else
-            {
+            } else {
                 std::cerr << "CutsceneRunner: unknown step '" << cmd << "'\n";
                 continue;
             }
@@ -129,8 +100,7 @@ bool CutsceneRunner::load(const std::string &path)
 
 // ---- playback ----
 
-void CutsceneRunner::start()
-{
+void CutsceneRunner::start() {
     currentStep = 0;
     finished = false;
     pendingMoves.clear();
@@ -142,8 +112,8 @@ bool CutsceneRunner::isFinished() const { return finished; }
 bool CutsceneRunner::isShowingDialogue() const { return inDialogue; }
 const std::string &CutsceneRunner::getId() const { return cutscene.id; }
 
-bool CutsceneRunner::update(World &world, GameUI &ui, bool confirmPressed, SoundManager &sound)
-{
+bool CutsceneRunner::update(World &world, GameUI &ui, bool confirmPressed,
+                            SoundManager &sound) {
     if (finished)
         return false;
 
@@ -152,13 +122,10 @@ bool CutsceneRunner::update(World &world, GameUI &ui, bool confirmPressed, Sound
         npc->updateAnimation();
 
     // If we're in a dialogue, wait for dismissal
-    if (inDialogue)
-    {
-        if (ui.updateTypewriter(confirmPressed))
-        {
+    if (inDialogue) {
+        if (ui.updateTypewriter(confirmPressed)) {
             sound.play(SoundEffect::select, ui.getRenderer().getWindow());
-            if (!ui.advanceDialogueLine())
-            {
+            if (!ui.advanceDialogueLine()) {
                 // Dialogue finished
                 inDialogue = false;
                 ++currentStep;
@@ -169,12 +136,10 @@ bool CutsceneRunner::update(World &world, GameUI &ui, bool confirmPressed, Sound
     }
 
     // If we're waiting frames
-    if (waitFrames > 0)
-    {
+    if (waitFrames > 0) {
         tickMovements(world);
         --waitFrames;
-        if (waitFrames <= 0)
-        {
+        if (waitFrames <= 0) {
             ++currentStep;
             processSteps(world, ui);
         }
@@ -182,12 +147,10 @@ bool CutsceneRunner::update(World &world, GameUI &ui, bool confirmPressed, Sound
     }
 
     // If we're syncing (waiting for all moves to finish)
-    if (currentStep < static_cast<int>(cutscene.steps.size()) &&
-        cutscene.steps[static_cast<size_t>(currentStep)].type == CutsceneStep::Type::sync)
-    {
+    if (currentStep < cutscene.steps.size() &&
+        cutscene.steps[currentStep].type == CutsceneStep::Type::sync) {
         tickMovements(world);
-        if (allMovesComplete(world))
-        {
+        if (allMovesComplete(world)) {
             pendingMoves.clear();
             ++currentStep;
             processSteps(world, ui);
@@ -200,30 +163,24 @@ bool CutsceneRunner::update(World &world, GameUI &ui, bool confirmPressed, Sound
     return !finished;
 }
 
-void CutsceneRunner::processSteps(World &world, GameUI &ui)
-{
+void CutsceneRunner::processSteps(World &world, GameUI &ui) {
     // Process immediate steps until we hit a blocking one or run out
-    while (currentStep < static_cast<int>(cutscene.steps.size()))
-    {
-        auto &step = cutscene.steps[static_cast<size_t>(currentStep)];
+    while (currentStep < cutscene.steps.size()) {
+        auto &step = cutscene.steps[currentStep];
 
-        switch (step.type)
-        {
-        case CutsceneStep::Type::move:
-        {
+        switch (step.type) {
+        case CutsceneStep::Type::move: {
             // Register a pending move (non-blocking)
             pendingMoves.push_back({step.target, {step.x, step.y}});
             ++currentStep;
             break;
         }
 
-        case CutsceneStep::Type::walk:
-        {
+        case CutsceneStep::Type::walk: {
             // Single tile step — register as a move to the adjacent tile
             Position cur = getEntityPosition(world, step.target);
             Position dest = cur;
-            switch (step.direction)
-            {
+            switch (step.direction) {
             case Direction::up:
                 dest.y--;
                 break;
@@ -242,24 +199,18 @@ void CutsceneRunner::processSteps(World &world, GameUI &ui)
             break;
         }
 
-        case CutsceneStep::Type::face:
-        {
+        case CutsceneStep::Type::face: {
             // Immediate: set facing
-            if (step.target == "player")
-            {
+            if (step.target == "player") {
                 world.getPlayer().setFacing(step.direction);
-            }
-            else
-            {
-                std::shared_ptr<NPC> npc = world.findNPCAt(world.getCurrentMapId(),
-                                                           getEntityPosition(world, step.target));
+            } else {
+                std::shared_ptr<NPC> npc =
+                    world.findNPCAt(world.getCurrentMapId(),
+                                    getEntityPosition(world, step.target));
                 // Also search by ID if position lookup fails
-                if (!npc)
-                {
-                    for (auto &n : world.getNPCs(world.getCurrentMapId()))
-                    {
-                        if (n->getId() == step.target)
-                        {
+                if (!npc) {
+                    for (auto &n : world.getNPCs(world.getCurrentMapId())) {
+                        if (n->getId() == step.target) {
                             npc = n;
                             break;
                         }
@@ -272,41 +223,34 @@ void CutsceneRunner::processSteps(World &world, GameUI &ui)
             break;
         }
 
-        case CutsceneStep::Type::say:
-        {
+        case CutsceneStep::Type::say: {
             // Blocking: show dialogue
             ui.startDialogue(step.speaker, step.lines);
             inDialogue = true;
             return; // Wait for dialogue dismissal
         }
 
-        case CutsceneStep::Type::wait:
-        {
+        case CutsceneStep::Type::wait: {
             // Blocking: wait N frames
             waitFrames = step.frames;
             return;
         }
 
-        case CutsceneStep::Type::sync:
-        {
+        case CutsceneStep::Type::sync: {
             // Blocking: wait for all pending moves to complete
             // The main update loop handles ticking and checking
             return;
         }
 
-        case CutsceneStep::Type::flag:
-        {
+        case CutsceneStep::Type::flag: {
             world.getPlayer().setFlag(step.flagName);
             ++currentStep;
             break;
         }
 
-        case CutsceneStep::Type::hide:
-        {
-            for (auto &n : world.getNPCs(world.getCurrentMapId()))
-            {
-                if (n->getId() == step.target)
-                {
+        case CutsceneStep::Type::hide: {
+            for (auto &n : world.getNPCs(world.getCurrentMapId())) {
+                if (n->getId() == step.target) {
                     n->setHidden(true);
                     break;
                 }
@@ -315,12 +259,9 @@ void CutsceneRunner::processSteps(World &world, GameUI &ui)
             break;
         }
 
-        case CutsceneStep::Type::show:
-        {
-            for (auto &n : world.getNPCs(world.getCurrentMapId()))
-            {
-                if (n->getId() == step.target)
-                {
+        case CutsceneStep::Type::show: {
+            for (auto &n : world.getNPCs(world.getCurrentMapId())) {
+                if (n->getId() == step.target) {
                     n->setHidden(false);
                     break;
                 }
@@ -337,34 +278,32 @@ void CutsceneRunner::processSteps(World &world, GameUI &ui)
 
 // ---- movement helpers ----
 
-Position CutsceneRunner::getEntityPosition(const World &world, const std::string &targetId) const
-{
+Position CutsceneRunner::getEntityPosition(const World &world,
+                                           const std::string &targetId) const {
     if (targetId == "player")
         return world.getPlayer().getPosition();
 
-    for (const auto &n : world.getNPCs(world.getCurrentMapId()))
-    {
+    for (const auto &n : world.getNPCs(world.getCurrentMapId())) {
         if (n->getId() == targetId)
             return n->getPosition();
     }
     return {0, 0};
 }
 
-bool CutsceneRunner::isEntityWalking(const World &world, const std::string &targetId) const
-{
+bool CutsceneRunner::isEntityWalking(const World &world,
+                                     const std::string &targetId) const {
     if (targetId == "player")
         return world.getPlayer().isMoving();
 
-    for (const auto &n : world.getNPCs(world.getCurrentMapId()))
-    {
+    for (const auto &n : world.getNPCs(world.getCurrentMapId())) {
         if (n->getId() == targetId)
             return n->isMoving();
     }
     return false;
 }
 
-void CutsceneRunner::stepEntityToward(World &world, const std::string &targetId, Position dest)
-{
+void CutsceneRunner::stepEntityToward(World &world, const std::string &targetId,
+                                      Position dest) {
     Position cur = getEntityPosition(world, targetId);
 
     // Decide direction: prefer X first, then Y
@@ -380,20 +319,15 @@ void CutsceneRunner::stepEntityToward(World &world, const std::string &targetId,
     else
         return; // already there
 
-    if (targetId == "player")
-    {
+    if (targetId == "player") {
         Player &player = world.getPlayer();
         Map &map = world.getMap(world.getCurrentMapId());
         map.setOccupied(player.getPosition(), false);
         player.move(dir);
         map.setOccupied(player.getPosition(), true);
-    }
-    else
-    {
-        for (auto &n : world.getNPCs(world.getCurrentMapId()))
-        {
-            if (n->getId() == targetId)
-            {
+    } else {
+        for (auto &n : world.getNPCs(world.getCurrentMapId())) {
+            if (n->getId() == targetId) {
                 n->setMoveDelay(24);
                 n->move(dir);
                 break;
@@ -402,19 +336,16 @@ void CutsceneRunner::stepEntityToward(World &world, const std::string &targetId,
     }
 }
 
-void CutsceneRunner::tickMovements(World &world)
-{
-    // For each entity, only process the FIRST pending move that isn't complete yet.
-    // This chains multiple moves for the same target sequentially.
+void CutsceneRunner::tickMovements(World &world) {
+    // For each entity, only process the FIRST pending move that isn't complete
+    // yet. This chains multiple moves for the same target sequentially.
     std::set<std::string> stepped; // entities we've already handled this tick
 
-    for (auto &pm : pendingMoves)
-    {
+    for (auto &pm : pendingMoves) {
         if (stepped.count(pm.targetId))
             continue; // Already processing an earlier move for this entity
 
-        if (isEntityWalking(world, pm.targetId))
-        {
+        if (isEntityWalking(world, pm.targetId)) {
             stepped.insert(pm.targetId); // Still animating, skip later moves
             continue;
         }
@@ -428,10 +359,8 @@ void CutsceneRunner::tickMovements(World &world)
     }
 }
 
-bool CutsceneRunner::allMovesComplete(const World &world) const
-{
-    for (const auto &pm : pendingMoves)
-    {
+bool CutsceneRunner::allMovesComplete(const World &world) const {
+    for (const auto &pm : pendingMoves) {
         if (isEntityWalking(world, pm.targetId))
             return false;
         Position cur = getEntityPosition(world, pm.targetId);
